[wip] extract fs methods and mock this module for tests
still not working
